+++ v3/pkg/application/webview_window_windows.go
@@ -279,6 +279,8 @@ func (w *windowsWebviewWindow) run() {
 
 	w.setResizable(!options.DisableResize)
 
+	w.setIgnoreMouseEvents(options.IgnoreMouseEvents)
+
 	if options.Frameless {
 		// Inform the application of the frame change this is needed to trigger the WM_NCCALCSIZE event.
 		// => https://learn.microsoft.com/en-us/windows/win32/dwm/customframe#removing-the-standard-frame
@@ -1760,4 +1762,19 @@ func (w *windowsWebviewWindow) setCloseButtonState(state ButtonState) {
 
 func (w *windowsWebviewWindow) setGWLStyle(style int) {
 	w32.SetWindowLong(w.hwnd, w32.GWL_STYLE, uint32(style))
+}
+
+func (w *windowsWebviewWindow) isIgnoreMouseEvents() bool {
+	exStyle := w32.GetWindowLong(w.hwnd, w32.GWL_EXSTYLE)
+	return exStyle&w32.WS_EX_TRANSPARENT != 0
+}
+
+func (w *windowsWebviewWindow) setIgnoreMouseEvents(ignore bool) {
+	exStyle := w32.GetWindowLong(w.hwnd, w32.GWL_EXSTYLE)
+	if ignore {
+		exStyle |= w32.WS_EX_LAYERED | w32.WS_EX_TRANSPARENT
+	} else {
+		exStyle &^= w32.WS_EX_TRANSPARENT
+	}
+	w32.SetWindowLong(w.hwnd, w32.GWL_EXSTYLE, uint32(exStyle))
 }